File Size script
So I've been fiddling with rotation and such to find rotation on GUI
elements (such as text boxes) is very different to that of 3D rotation
so I modified the IP text box script to make something similar for file
size. I imagine I can add to with other things we need to check for (for
different levels) but cross that bridge when we get there. What I was
thinking is instead of right-clicking and rotating was maybe just
rolling the mouse over the text and getting a popup box. The idea for it
is here:
http://answers.unity3d.com/questions/17812/text-box-pop-up-on-mouseover.html
If you want to have a go, go ahead. If not upload a more up-to-date
version of the game and I'll have a go.