Read Description
This is going to be a bit of a read. Sorry!

First things first the bad IPs are drawn on screen. And the timer is now
a global variable so we can use it for things. I've already used it to
destroy the boxes at the end of the level.

But the main thing I've done is create the other levels. And the boxes
hold their own IP address and the other data if you're not on the first
level. All that's left is to fill in what that data is. If anyone wants
to do that because I know we had some joke ideas, feel free! Otherwise
I'll do it later. If you do it then out of the 20, the last 3 are the
virus options. Feel free to change that too if you like. And what the
data is for each level because I just took the ideas we had written down
on trello.

Steven, I put the code for the other data in the original IPspawn. It
was just so the levels all use the same code. I did leave IPspawnV2
though so I didn't delete the raycasting code. Which should also either
be moved somewhere else (or turn that script into the raycasting
script). I hope that's all ok :)

I have also update the read me (which you will find on the home page).
PLEASE READ IT!!!!! It has some details about how additions should fit
with the code we have. As well as more detail about what's left. I also
added a few notes into the code I was working with so we don't get lost.