Slight Progress
Ok so right now if you directly right click on the box in SH_test then
it will cycle between the blank and the file size which stays the same
which is half way there. The blank IP part I think comes from line 36
"address = "";", I think it resets the string each time. I tried taking
it out but it made it worse by start the test level with all the IPs
printed in one string. If you right click outside the box then it will
cycle between 2 debug messages and stay like that even if you right
click on the box afterwards. I'm gonna email Andy tonight see if I can
get his help and I'll try and fix if he responds. I'll try to do a bit
more of the write up where I can as I know I haven't done as much for
the game as others.