Almost finished
The camera and menus need doing. And swapping between the four levels.
The data for some of the levels needs to be made up, you'll find that in
IPspawn if you want. I tried creating a popup if a virus makes it
through, it doesn't work though. You'll probably have to change where
everything is being made when you change the camera.